[GREEN-53] refactor/#3 : convert package of Address to common.vo
